requirements.txt
click==7.1.2
Flask==1.1.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.2.0
pytesseract==0.3.4
python-markdown-math==0.7
Werkzeug==1.0.1
zipp==3.1.0
